created useCategories custom hook
